refactor(icons): move URL icon helpers into utils
Give custom icons a dedicated `@icons` alias and move URL-to-icon helpers into the shared utils module so the remaining libs/icons barrel can be removed cleanly. This keeps custom icon imports explicit while reducing barrel indirection ahead of the broader import-graph cleanup.

Made-with: Cursor
