+++ src/components/molecules/PostHashtags/PostHashtags.tsx
@@ -2,20 +2,20 @@ import { JSX } from 'react';
 import { RemarkAnchorProps } from '../PostText/PostText.types';
 import { extractTextFromChildren } from '../PostText/PostText.utils';
 import * as Libs from '@/libs';
-import * as Icons from '@/libs/icons';
 import * as Atoms from '@/atoms';
+import { Bitkit, Blocktank, BTCIcon, PubkyIcon, Synonym, Tether } from '@icons';
 
 type TagIcons = {
   [key: string]: JSX.Element | undefined;
 };
 
 const tagIcons: TagIcons = {
-  '#synonym': <Icons.Synonym size={24} />,
-  '#tether': <Icons.Tether size={24} />,
-  '#pubky': <Icons.PubkyIcon size={22} />,
-  '#bitkit': <Icons.Bitkit size={24} />,
-  '#blocktank': <Icons.Blocktank size={24} />,
-  '#bitcoin': <Icons.BTCIcon size={24} />,
+  '#synonym': <Synonym size={24} />,
+  '#tether': <Tether size={24} />,
+  '#pubky': <PubkyIcon size={22} />,
+  '#bitkit': <Bitkit size={24} />,
+  '#blocktank': <Blocktank size={24} />,
+  '#bitcoin': <BTCIcon size={24} />,
 };
 
 export const PostHashtags = (props: RemarkAnchorProps) => {
