+++ src/components/molecules/ProfilePageFilterBar/ProfilePageFilterBar.tsx
@@ -8,7 +8,7 @@ import * as Hooks from '@/hooks';
 import * as Types from '@/app/profile/types';
 import * as Config from '@/config';
 import { Bell, StickyNote, MessageCircle, UsersRound, HeartHandshake, Tag } from 'lucide-react';
-import { UsersRound2 } from '@/libs/icons';
+import { UsersRound2 } from '@icons';
 export interface ProfilePageFilterBarItem {
   icon: React.ComponentType<{
     className?: string;
