+++ src/components/molecules/ShowMoreReplies/ShowMoreReplies.tsx
@@ -4,7 +4,7 @@ import * as Atoms from '@/atoms';
 import * as Libs from '@/libs';
 import { useTranslations } from 'next-intl';
 import { CirclePlus } from 'lucide-react';
-import { RoundedCorner } from '@/libs/icons';
+import { RoundedCorner } from '@icons';
 interface ShowMoreRepliesProps {
   /** Number of remaining replies to show */
   count: number;
