+++ src/components/organisms/HumanInviteCode/HumanInviteCode.tsx
@@ -17,7 +17,7 @@ import type { HumanInviteCodeProps } from './HumanInviteCode.types';
  * @param onSuccess - Called with the trimmed invite code; may be async. Throws on validation failure.
  */
 import { CircleCheck, Server, ArrowLeft, Loader2, ArrowRight } from 'lucide-react';
-import { XTwitter, Telegram } from '@/libs/icons';
+import { XTwitter, Telegram } from '@icons';
 export const HumanInviteCode = ({ onBack, onSuccess }: HumanInviteCodeProps) => {
   const t = useTranslations('onboarding.inviteCode');
   const tCommon = useTranslations('common');
