--- src/libs/icons/index.ts
@@ -1,2 +1,0 @@
-export * from './icons';
-export * from './urlToIcon';
