+++ src/libs/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './utils';
 export * from './utils.constants';
+export * from './urlToIcon';
 export type * from './utils.types';
